converted http-status-codes enum values to type union
This template-literal method returns a union type of the enum values as strings. This won’t be suitable for the StatusCodes enum since the values are numbers. I can’t prevent this, so I'll have to adapt to it.

This answer on Stack Overflow tries something complex but is overkill for my use case. https://stackoverflow.com/a/69090186/15350139

squash